--- src/schema.js
@@ -1,54 +1,0 @@
-export const SCHEMA_MAP = [
-  {
-    path: "/eo-missions/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/eo-missions/parent.json",
-  },
-  {
-    path: "/eo-missions/<id>/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/eo-missions/children.json",
-  },
-  {
-    path: "/products/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/products/parent.json",
-  },
-  {
-    path: "/products/<id>/collection.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/products/children.json",
-  },
-  {
-    path: "/projects/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/projects/parent.json",
-  },
-  {
-    path: "/projects/<id>/collection.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/projects/children.json",
-  },
-  {
-    path: "/themes/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/themes/parent.json",
-  },
-  {
-    path: "/themes/<id>/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/themes/children.json",
-  },
-  {
-    path: "/variables/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/variables/parent.json",
-  },
-  {
-    path: "/variables/<id>/catalog.json",
-    url: "https://esa-earthcode.github.io/open-science-catalog-validation/schemas/variables/children.json",
-  },
-];
-
-export function getSchemaURL(inputPath) {
-  for (const schema of SCHEMA_MAP) {
-    const pathPattern = schema.path.replace(/<id>/g, "[^/]+");
-    const regex = new RegExp(`^${pathPattern}$`);
-
-    if (regex.test(inputPath)) {
-      return schema.url;
-    }
-  }
-  return null;
-}
